Added Kyocera FS-1118MFP
This printer works "excellent" with Linux (according to Linux foundation). I can confirm it works with any Mac OS version (tested with 10.10, 10.11, 10.12 and 10.13). On Mac OS X it is possible to utilize a PPD (printer description format), and this seamlessly works with past and future operating systems. Materials (drums and toners) are very cheap as well. https://ebay.us/sYyXnR