Added ASUS ROG Strix G15 AMD Ryzen 9 6900HX.
The ASUS ROG Strix G15 G513RM-HF150W Eclipse Gray, Ryzen 9 6900HX, 16GB RAM, 1TB SSD, GeForce RTX 3060 has been confirmed to play nicely with Linux, so I added it to the list. This is one of the suggested devices if you are planning to work on graphics, UX or game development with Rust on your Linux machine. https://amzn.to/3rh2oG1